MSF-89 - Initialization flow - Initial separation of responsibilities.
