+++ Factory/src/main/kotlin/net/milosvasic/factory/mail/execution/flow/Flow.kt
@@ -6,7 +6,5 @@ interface Flow<T, M, D> : Runnable {
 
     fun width(subject: T): Flow<T, M, D>
 
-    fun perform(what: M): Flow<T, M, D>
-
     fun onFinish(callback: FlowCallback<D>): Flow<T, M, D>
 }
