+++ Factory/src/main/kotlin/net/milosvasic/factory/mail/execution/flow/FlowChain.kt
@@ -1,0 +1,6 @@
+package net.milosvasic.factory.mail.execution.flow
+
+interface FlowChain {
+
+    fun chain(flow: Flow<*, *, *>): FlowChain
+}
