+++ Factory/src/main/kotlin/net/milosvasic/factory/mail/execution/flow/FlowPerform.kt
@@ -1,0 +1,6 @@
+package net.milosvasic.factory.mail.execution.flow
+
+interface FlowPerform<T, M, D> {
+
+    fun perform(what: M): Flow<T, M, D>
+}
